rebalance controller test suit runtime
split tests in more files to rebalance runtime for each parallel test suite

Reference: CHECK-2620
