+++ test/controllers/elastic_search_2_test.rb
@@ -233,5 +233,97 @@ class ElasticSearch2Test < ActionController::TestCase
     end
   end
 
+  [:asc, :desc].each do |order|
+    test "should sort by item title #{order}" do
+      RequestStore.store[:skip_cached_field_update] = false
+      pender_url = CheckConfig.get('pender_url_private') + '/api/medias'
+      url = 'http://test.com'
+      response = '{"type":"media","data":{"url":"' + url + '/normalized","type":"item", "title": "b-item"}}'
+      WebMock.stub_request(:get, pender_url).with({ query: { url: url } }).to_return(body: response)
+      l = create_media(account: create_valid_account, url: url)
+      i = create_uploaded_image file: 'c-item.png'
+      v = create_uploaded_video file: 'd-item.mp4'
+      a = create_uploaded_audio file: 'e-item.mp3'
+      t = create_team
+      p = create_project team: t
+      pm1 = create_project_media project: p, quote: 'a-item', disable_es_callbacks: false
+      pm2 = create_project_media project: p, media: l, disable_es_callbacks: false
+      pm3 = create_project_media project: p, media: i, disable_es_callbacks: false
+      pm3.analysis = { file_title: 'c-item' }; pm3.save
+      pm4 = create_project_media project: p, media: v, disable_es_callbacks: false
+      pm4.analysis = { file_title: 'd-item' }; pm4.save
+      pm5 = create_project_media project: p, media: a, disable_es_callbacks: false
+      pm5.analysis = { file_title: 'e-item' }; pm5.save
+      sleep 2
+      orders = {asc: [pm1, pm2, pm3, pm4, pm5], desc: [pm5, pm4, pm3, pm2, pm1]}
+      query = { projects: [p.id], keyword: 'item', sort: 'title', sort_type: order.to_s }
+      result = CheckSearch.new(query.to_json, nil, t.id)
+      assert_equal 5, result.medias.count
+      assert_equal orders[order.to_sym].map(&:id), result.medias.map(&:id)
+      query = { projects: [p.id], sort: 'title', sort_type: order.to_s }
+      result = CheckSearch.new(query.to_json, nil, t.id)
+      assert_equal 5, result.medias.count
+      assert_equal orders[order.to_sym].map(&:id), result.medias.map(&:id)
+      # update analysis
+      pm3.analysis = { file_title: 'f-item' }
+      pm6 = create_project_media project: p, quote: 'DUPPER-item', disable_es_callbacks: false
+      sleep 2
+      orders = {asc: [pm1, pm2, pm4, pm6, pm5, pm3], desc: [pm3, pm5, pm6, pm4, pm2, pm1]}
+      result = CheckSearch.new(query.to_json, nil, t.id)
+      assert_equal 6, result.medias.count
+      assert_equal orders[order.to_sym].map(&:id), result.medias.map(&:id)
+    end
+  end
+
+  test "should search by source" do
+    t = create_team
+    s = create_source team: t
+    s2 = create_source team: t
+    s3 = create_source team: t
+    pm = create_project_media team: t, source: s, disable_es_callbacks: false, skip_autocreate_source: false
+    pm2 = create_project_media team: t, source: s, disable_es_callbacks: false, skip_autocreate_source: false
+    pm3 = create_project_media team: t, source: s2, disable_es_callbacks: false, skip_autocreate_source: false
+    sleep 2
+    result = CheckSearch.new({ sources: [s.id] }.to_json, nil, t.id)
+    assert_equal [pm.id, pm2.id], result.medias.map(&:id).sort
+    result = CheckSearch.new({ sources: [s2.id] }.to_json, nil, t.id)
+    assert_equal [pm3.id], result.medias.map(&:id)
+    result = CheckSearch.new({ sources: [s.id, s2.id] }.to_json, nil, t.id)
+    assert_equal [pm.id, pm2.id, pm3.id], result.medias.map(&:id).sort
+    result = CheckSearch.new({ sources: [s3.id] }.to_json, nil, t.id)
+    assert_empty result.medias
+    result = CheckSearch.new({ sources: [s2.id], show: ['links'] }.to_json, nil, t.id)
+    assert_equal [pm3.id], result.medias.map(&:id)
+  end
+
+  test "should search trash and unconfirmed items" do
+    t = create_team
+    pm = create_project_media team: t, disable_es_callbacks: false
+    pm2 = create_project_media team: t, archived: CheckArchivedFlags::FlagCodes::TRASHED, disable_es_callbacks: false
+    pm3 = create_project_media team: t, archived: CheckArchivedFlags::FlagCodes::TRASHED, disable_es_callbacks: false
+    pm4 = create_project_media team: t, archived: CheckArchivedFlags::FlagCodes::UNCONFIRMED, disable_es_callbacks: false
+    sleep 2
+    assert_equal [pm2, pm3], pm.check_search_trash.medias.sort
+    assert_equal [pm, pm4], t.check_search_team.medias.sort
+  end
+
+  test "should adjust ES window size" do
+    t = create_team
+    u = create_user
+    pm = create_project_media quote: 'claim a', disable_es_callbacks: false
+    sleep 2
+    create_team_user team: t, user: u, role: 'admin'
+    with_current_user_and_team(u ,t) do
+      assert_nothing_raised do
+        query = 'query Search { search(query: "{\"keyword\":\"claim\",\"eslimit\":20000,\"esoffset\":0}") {medias(first:20){edges{node{dbid}}}}}'
+        post :create, params: { query: query }
+        assert_response :success
+        query = 'query Search { search(query: "{\"keyword\":\"claim\",\"eslimit\":10000,\"esoffset\":20}") {medias(first:20){edges{node{dbid}}}}}'
+        post :create, params: { query: query }
+        assert_response :success
+      end
+    end
+  end
+
   # Please add new tests to test/controllers/elastic_search_7_test.rb
 end
