+++ test/controllers/elastic_search_3_test.rb
@@ -149,5 +149,146 @@ class ElasticSearch3Test < ActionController::TestCase
     assert_equal [pm.id, pm2.id].sort, result.medias.map(&:id).sort
   end
 
+  test "should sort by cluster_published_reports_count" do
+    RequestStore.store[:skip_cached_field_update] = false
+    t = create_team
+    t2 = create_team
+    f = create_feed
+    f.teams << t
+    f.teams << t2
+    FeedTeam.update_all(shared: true)
+    pm1 = create_project_media team: t
+    c1 = create_cluster project_media: pm1
+    c1.project_medias << pm1
+    pm2 = create_project_media team: t
+    c2 = create_cluster project_media: pm2
+    c2.project_medias << pm2
+    pm2_t2 = create_project_media team: t2
+    c1.project_medias << pm2_t2
+    publish_report(pm2)
+    publish_report(pm2_t2)
+    publish_report(pm1)
+    sleep 2
+    Team.stubs(:current).returns(t)
+    query = { clusterize: true, feed_id: f.id, sort: 'cluster_published_reports_count' }
+    result = CheckSearch.new(query.to_json)
+    assert_equal [pm1.id, pm2.id], result.medias.map(&:id)
+    query[:sort_type] = 'asc'
+    result = CheckSearch.new(query.to_json)
+    assert_equal [pm2.id, pm1.id], result.medias.map(&:id)
+    # filter by published_by filter `cluster_published_reports`
+    query = { clusterize: true, feed_id: f.id, cluster_published_reports: [t.id, t2.id]}
+    result = CheckSearch.new(query.to_json)
+    assert_equal [pm1.id, pm2.id], result.medias.map(&:id).sort
+    query = { clusterize: true, feed_id: f.id, cluster_published_reports: [t2.id]}
+    result = CheckSearch.new(query.to_json)
+    assert_equal [pm1.id], result.medias.map(&:id)
+    Team.unstub(:current)
+  end
+
+  test "should sort by cluster_first_item_at" do
+    t = create_team
+    f = create_feed
+    f.teams << t
+    FeedTeam.update_all(shared: true)
+    Time.stubs(:now).returns(Time.new - 2.week)
+    pm1 = create_project_media team: t
+    c1 = create_cluster project_media: pm1
+    c1.project_medias << pm1
+    Time.stubs(:now).returns(Time.new - 1.week)
+    pm2 = create_project_media team: t
+    c2 = create_cluster project_media: pm2
+    c2.project_medias << pm2
+    Time.stubs(:now).returns(Time.new - 3.week)
+    pm3 = create_project_media team: t
+    c3 = create_cluster project_media: pm3
+    c3.project_medias << pm3
+    Time.unstub(:now)
+    sleep 2
+    Team.stubs(:current).returns(t)
+    query = { clusterize: true, feed_id: f.id, sort: 'cluster_first_item_at' }
+    result = CheckSearch.new(query.to_json)
+    assert_equal [pm2.id, pm1.id, pm3.id], result.medias.map(&:id)
+    query[:sort_type] = 'asc'
+    result = CheckSearch.new(query.to_json)
+    assert_equal [pm3.id, pm1.id, pm2.id], result.medias.map(&:id)
+    Team.unstub(:current)
+  end
+
+  test "should sort by clusters requests count" do
+    RequestStore.store[:skip_cached_field_update] = false
+    create_annotation_type_and_fields('Smooch', { 'Data' => ['JSON', false] })
+    t = create_team
+    f = create_feed
+    f.teams << t
+    FeedTeam.update_all(shared: true)
+    u = create_user
+    create_team_user team: t, user: u, role: 'admin'
+    pm1 = create_project_media team: t
+    pm1_1 = create_project_media team: t
+    pm2 = create_project_media team: t
+    create_dynamic_annotation annotation_type: 'smooch', annotated: pm1
+    create_dynamic_annotation annotation_type: 'smooch', annotated: pm2
+    c1 = create_cluster project_media: pm1
+    c2 = create_cluster project_media: pm2
+    c1.project_medias << pm1
+    c1.project_medias << pm1_1
+    c2.project_medias << pm2
+    sleep 2
+    with_current_user_and_team(u, t) do
+      create_dynamic_annotation annotation_type: 'smooch', annotated: pm1
+      create_dynamic_annotation annotation_type: 'smooch', annotated: pm1_1
+      sleep 2
+      es1 = $repository.find(get_es_id(pm1))
+      es2 = $repository.find(get_es_id(pm2))
+      assert_equal c1.requests_count(true), es1['cluster_requests_count']
+      assert_equal c2.requests_count(true), es2['cluster_requests_count']
+      query = { clusterize: true, feed_id: f.id, sort: 'cluster_requests_count' }
+      result = CheckSearch.new(query.to_json)
+      assert_equal [pm1.id, pm2.id], result.medias.map(&:id)
+      query[:sort_type] = 'asc'
+      result = CheckSearch.new(query.to_json)
+      assert_equal [pm2.id, pm1.id], result.medias.map(&:id)
+    end
+  end
+
+  test "should sort by cluster_size" do
+    t = create_team
+    f = create_feed 
+    f.teams << t
+    FeedTeam.update_all(shared: true)
+    u = create_user
+    create_team_user team: t, user: u, role: 'admin'
+    pm1 = create_project_media team: t
+    pm1_1 = create_project_media team: t
+    pm1_2 = create_project_media team: t
+    pm2 = create_project_media team: t
+    pm2_1 = create_project_media team: t
+    pm2_2 = create_project_media team: t
+    pm2_3 = create_project_media team: t
+    pm3 = create_project_media team: t
+    pm3_1 = create_project_media team: t
+    c1 = create_cluster project_media: pm1
+    c2 = create_cluster project_media: pm2
+    c3 = create_cluster project_media: pm3
+    c1.project_medias << pm1
+    c1.project_medias << pm1_1
+    c1.project_medias << pm1_2
+    c2.project_medias << pm2
+    c2.project_medias << pm2_1
+    c2.project_medias << pm2_2
+    c2.project_medias << pm2_3
+    c3.project_medias << pm3
+    c3.project_medias << pm3_1
+    sleep 2
+    with_current_user_and_team(u, t) do
+      query = { clusterize: true, feed_id: f.id, sort: 'cluster_size' }
+      result = CheckSearch.new(query.to_json)
+      assert_equal [pm2.id, pm1.id, pm3.id], result.medias.map(&:id)
+      query[:sort_type] = 'asc'
+      result = CheckSearch.new(query.to_json)
+      assert_equal [pm3.id, pm1.id, pm2.id], result.medias.map(&:id)
+    end
+  end
   # Please add new tests to test/controllers/elastic_search_7_test.rb
 end
