+++ test/controllers/elastic_search_6_test.rb
@@ -6,48 +6,6 @@ class ElasticSearch6Test < ActionController::TestCase
     setup_elasticsearch
   end
 
-  [:asc, :desc].each do |order|
-    test "should sort by item title #{order}" do
-      RequestStore.store[:skip_cached_field_update] = false
-      pender_url = CheckConfig.get('pender_url_private') + '/api/medias'
-      url = 'http://test.com'
-      response = '{"type":"media","data":{"url":"' + url + '/normalized","type":"item", "title": "b-item"}}'
-      WebMock.stub_request(:get, pender_url).with({ query: { url: url } }).to_return(body: response)
-      l = create_media(account: create_valid_account, url: url)
-      i = create_uploaded_image file: 'c-item.png'
-      v = create_uploaded_video file: 'd-item.mp4'
-      a = create_uploaded_audio file: 'e-item.mp3'
-      t = create_team
-      p = create_project team: t
-      pm1 = create_project_media project: p, quote: 'a-item', disable_es_callbacks: false
-      pm2 = create_project_media project: p, media: l, disable_es_callbacks: false
-      pm3 = create_project_media project: p, media: i, disable_es_callbacks: false
-      pm3.analysis = { file_title: 'c-item' }; pm3.save
-      pm4 = create_project_media project: p, media: v, disable_es_callbacks: false
-      pm4.analysis = { file_title: 'd-item' }; pm4.save
-      pm5 = create_project_media project: p, media: a, disable_es_callbacks: false
-      pm5.analysis = { file_title: 'e-item' }; pm5.save
-      sleep 2
-      orders = {asc: [pm1, pm2, pm3, pm4, pm5], desc: [pm5, pm4, pm3, pm2, pm1]}
-      query = { projects: [p.id], keyword: 'item', sort: 'title', sort_type: order.to_s }
-      result = CheckSearch.new(query.to_json, nil, t.id)
-      assert_equal 5, result.medias.count
-      assert_equal orders[order.to_sym].map(&:id), result.medias.map(&:id)
-      query = { projects: [p.id], sort: 'title', sort_type: order.to_s }
-      result = CheckSearch.new(query.to_json, nil, t.id)
-      assert_equal 5, result.medias.count
-      assert_equal orders[order.to_sym].map(&:id), result.medias.map(&:id)
-      # update analysis
-      pm3.analysis = { file_title: 'f-item' }
-      pm6 = create_project_media project: p, quote: 'DUPPER-item', disable_es_callbacks: false
-      sleep 2
-      orders = {asc: [pm1, pm2, pm4, pm6, pm5, pm3], desc: [pm3, pm5, pm6, pm4, pm2, pm1]}
-      result = CheckSearch.new(query.to_json, nil, t.id)
-      assert_equal 6, result.medias.count
-      assert_equal orders[order.to_sym].map(&:id), result.medias.map(&:id)
-    end
-  end
-
   [:created_at, :updated_at, :last_seen].each do |field|
     test "should filter by #{field} range" do
       RequestStore.store[:skip_cached_field_update] = false
