+++ test/controllers/elastic_search_9_test.rb
@@ -246,139 +246,5 @@ class ElasticSearch9Test < ActionController::TestCase
     end
   end
 
-  test "should search by source" do
-    t = create_team
-    s = create_source team: t
-    s2 = create_source team: t
-    s3 = create_source team: t
-    pm = create_project_media team: t, source: s, disable_es_callbacks: false, skip_autocreate_source: false
-    pm2 = create_project_media team: t, source: s, disable_es_callbacks: false, skip_autocreate_source: false
-    pm3 = create_project_media team: t, source: s2, disable_es_callbacks: false, skip_autocreate_source: false
-    sleep 2
-    result = CheckSearch.new({ sources: [s.id] }.to_json, nil, t.id)
-    assert_equal [pm.id, pm2.id], result.medias.map(&:id).sort
-    result = CheckSearch.new({ sources: [s2.id] }.to_json, nil, t.id)
-    assert_equal [pm3.id], result.medias.map(&:id)
-    result = CheckSearch.new({ sources: [s.id, s2.id] }.to_json, nil, t.id)
-    assert_equal [pm.id, pm2.id, pm3.id], result.medias.map(&:id).sort
-    result = CheckSearch.new({ sources: [s3.id] }.to_json, nil, t.id)
-    assert_empty result.medias
-    result = CheckSearch.new({ sources: [s2.id], show: ['links'] }.to_json, nil, t.id)
-    assert_equal [pm3.id], result.medias.map(&:id)
-  end
-
-  test "should search trash and unconfirmed items" do
-    t = create_team
-    pm = create_project_media team: t, disable_es_callbacks: false
-    pm2 = create_project_media team: t, archived: CheckArchivedFlags::FlagCodes::TRASHED, disable_es_callbacks: false
-    pm3 = create_project_media team: t, archived: CheckArchivedFlags::FlagCodes::TRASHED, disable_es_callbacks: false
-    pm4 = create_project_media team: t, archived: CheckArchivedFlags::FlagCodes::UNCONFIRMED, disable_es_callbacks: false
-    sleep 2
-    assert_equal [pm2, pm3], pm.check_search_trash.medias.sort
-    assert_equal [pm, pm4], t.check_search_team.medias.sort
-  end
-
-  test "should adjust ES window size" do
-    t = create_team
-    u = create_user
-    pm = create_project_media quote: 'claim a', disable_es_callbacks: false
-    sleep 2
-    create_team_user team: t, user: u, role: 'admin'
-    with_current_user_and_team(u ,t) do
-      assert_nothing_raised do
-        query = 'query Search { search(query: "{\"keyword\":\"claim\",\"eslimit\":20000,\"esoffset\":0}") {medias(first:20){edges{node{dbid}}}}}'
-        post :create, params: { query: query }
-        assert_response :success
-        query = 'query Search { search(query: "{\"keyword\":\"claim\",\"eslimit\":10000,\"esoffset\":20}") {medias(first:20){edges{node{dbid}}}}}'
-        post :create, params: { query: query }
-        assert_response :success
-      end
-    end
-  end
-
-  test "should search by media url" do
-    url = 'http://test.com'
-    pender_url = CheckConfig.get('pender_url_private') + '/api/medias'
-    response = '{"type":"media","data":{"url":"' + url + '","type":"item","title": "media_title"}}'
-    WebMock.stub_request(:get, pender_url).with({ query: { url: url } }).to_return(body: response)
-    t = create_team
-    u = create_user
-    create_team_user team: t, user: u, role: 'admin'
-    with_current_user_and_team(u ,t) do
-      pm = create_project_media team: t, url: url, disable_es_callbacks: false
-      sleep 2
-      result = $repository.find(get_es_id(pm))['url']
-      assert_equal result, url
-      results = CheckSearch.new({ keyword: 'test.com' }.to_json)
-      assert_equal [pm.id], results.medias.map(&:id)
-      results = CheckSearch.new({ keyword: 'test2.com' }.to_json)
-      assert_empty results.medias.map(&:id)
-      results = CheckSearch.new({keyword: 'test.com', keyword_fields: {fields: ['url']}}.to_json)
-      assert_equal [pm.id], results.medias.map(&:id)
-      results = CheckSearch.new({keyword: 'test.com', keyword_fields: {fields: ['title']}}.to_json)
-      assert_empty results.medias.map(&:id)
-    end
-  end
-
-  test "should filter items by channel" do
-    t = create_team
-    u = create_user
-    create_team_user team: t, user: u, role: 'admin'
-    with_current_user_and_team(u ,t) do
-      pm = create_project_media team: t, quote: 'claim a', channel: { main: CheckChannels::ChannelCodes::MANUAL }, disable_es_callbacks: false
-      pm2 = create_project_media team: t, channel: { main: CheckChannels::ChannelCodes::FETCH }, disable_es_callbacks: false
-      pm3 = create_project_media team: t, channel: { main: CheckChannels::ChannelCodes::API }, disable_es_callbacks: false
-      pm4 = create_project_media team: t, quote: 'claim b', channel: { main: CheckChannels::ChannelCodes::ZAPIER }, disable_es_callbacks: false
-      # tipline items
-      pm5 = create_project_media team: t, channel: { main: CheckChannels::ChannelCodes::WHATSAPP }, disable_es_callbacks: false
-      pm6 = create_project_media team: t, channel: { main: CheckChannels::ChannelCodes::MESSENGER }, disable_es_callbacks: false
-      pm7 = create_project_media team: t, channel: { main: CheckChannels::ChannelCodes::TWITTER }, disable_es_callbacks: false
-      pm8 = create_project_media team: t, channel: { main: CheckChannels::ChannelCodes::TELEGRAM }, disable_es_callbacks: false
-      pm9 = create_project_media team: t, channel: { main: CheckChannels::ChannelCodes::VIBER }, disable_es_callbacks: false
-      pm10 = create_project_media team: t, channel: { main: CheckChannels::ChannelCodes::LINE }, disable_es_callbacks: false
-      tipline_ids = [pm5.id, pm6.id, pm7.id, pm8.id, pm9.id, pm10.id]
-      sleep 2
-      # Hit PG
-      results = CheckSearch.new({ channels: [CheckChannels::ChannelCodes::MANUAL] }.to_json)
-      assert_equal [pm.id], results.medias.map(&:id)
-      results = CheckSearch.new({ channels: [CheckChannels::ChannelCodes::MANUAL, CheckChannels::ChannelCodes::API] }.to_json)
-      assert_equal [pm.id, pm3.id], results.medias.map(&:id).sort
-      # Hit ES
-      results = CheckSearch.new({ keyword: 'claim', channels: [CheckChannels::ChannelCodes::MANUAL, CheckChannels::ChannelCodes::API] }.to_json)
-      assert_equal [pm.id], results.medias.map(&:id)
-      # filter by any tipline
-      results = CheckSearch.new({ channels: ['any_tipline'] }.to_json)
-      assert_equal tipline_ids, results.medias.map(&:id).sort
-      results = CheckSearch.new({ channels: ['any_tipline', CheckChannels::ChannelCodes::MANUAL, CheckChannels::ChannelCodes::TWITTER] }.to_json)
-      assert_equal tipline_ids.concat([pm.id]).sort, results.medias.map(&:id).sort
-    end
-  end
-
-  test "should filter items by channel in main and others" do
-    t = create_team
-    u = create_user
-    create_team_user team: t, user: u, role: 'admin'
-    with_current_user_and_team(u ,t) do
-      pm = create_project_media team: t, quote: 'claim a', channel: { main: CheckChannels::ChannelCodes::MANUAL }, disable_es_callbacks: false
-      pm2 = create_project_media team: t, quote: 'claim b', channel: { main: CheckChannels::ChannelCodes::ZAPIER }, disable_es_callbacks: false
-      # tipline items
-      pm3 = create_project_media team: t, channel: { main: CheckChannels::ChannelCodes::WHATSAPP }, disable_es_callbacks: false
-      pm.channel = { main: CheckChannels::ChannelCodes::MANUAL, others: [CheckChannels::ChannelCodes::WHATSAPP, CheckChannels::ChannelCodes::MESSENGER] }
-      pm.save!
-      sleep 2
-      results = CheckSearch.new({ channels: [CheckChannels::ChannelCodes::MANUAL] }.to_json)
-      assert_equal [pm.id], results.medias.map(&:id)
-      results = CheckSearch.new({ channels: [CheckChannels::ChannelCodes::WHATSAPP] }.to_json)
-      assert_equal [pm.id, pm3.id], results.medias.map(&:id).sort
-      results = CheckSearch.new({ channels: [CheckChannels::ChannelCodes::WHATSAPP, CheckChannels::ChannelCodes::ZAPIER] }.to_json)
-      assert_equal [pm.id, pm2.id, pm3.id], results.medias.map(&:id).sort
-      results = CheckSearch.new({ channels: [CheckChannels::ChannelCodes::MESSENGER] }.to_json)
-      assert_equal [pm.id], results.medias.map(&:id)
-      # filter by any tipline
-      results = CheckSearch.new({ channels: ['any_tipline'] }.to_json)
-      assert_equal [pm.id, pm3.id], results.medias.map(&:id).sort
-    end
-  end
-
   # Please add new tests to test/controllers/elastic_search_10_test.rb
 end
