+++ test/controllers/graphql_controller_3_test.rb
@@ -920,684 +920,4 @@ class GraphqlController3Test < ActionController::TestCase
     assert_nil JSON.parse(@response.body)['errors']
   end
 
-  test "should get nested comment" do
-    u = create_user is_admin: true
-    t = create_team
-    p = create_project team: t
-    pm = create_project_media project: p
-    c1 = create_comment annotated: pm, text: 'Parent'
-    c2 = create_comment annotated: c1, text: 'Child'
-    authenticate_with_user(u)
-    query = %{
-      query {
-        project_media(ids: "#{pm.id},#{p.id}") {
-          comments: annotations(first: 10000, annotation_type: "comment") {
-            edges {
-              node {
-                ... on Comment {
-                  id
-                  text
-                  comments: annotations(first: 10000, annotation_type: "comment") {
-                    edges {
-                      node {
-                        ... on Comment {
-                          id
-                          text
-                        }
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    comments = JSON.parse(@response.body)['data']['project_media']['comments']['edges']
-    assert_equal 1, comments.size
-    assert_equal 'Parent', comments[0]['node']['text']
-    child_comments = comments[0]['node']['comments']['edges']
-    assert_equal 1, child_comments.size
-    assert_equal 'Child', child_comments[0]['node']['text']
-  end
-
-  test "should create and retrieve clips" do
-    json_schema = {
-      type: 'object',
-      required: ['label'],
-      properties: {
-        label: { type: 'string' }
-      }
-    }
-    DynamicAnnotation::AnnotationType.reset_column_information
-    create_annotation_type_and_fields('Clip', {}, json_schema)
-    u = create_user is_admin: true
-    p = create_project
-    pm = create_project_media project: p
-    authenticate_with_user(u)
-
-    query = 'mutation { createDynamic(input: { annotation_type: "clip", annotated_type: "ProjectMedia", annotated_id: "' + pm.id.to_s + '", fragment: "t=10,20", set_fields: "{\"label\":\"Clip Label\"}" }) { dynamic { data, parsed_fragment } } }'
-    assert_difference 'Annotation.where(annotation_type: "clip").count', 1 do
-      post :create, params: { query: query, team: pm.team.slug }
-    end
-    assert_response :success
-    annotation = JSON.parse(@response.body)['data']['createDynamic']['dynamic']
-    assert_equal 'Clip Label', annotation['data']['label']
-    assert_equal({ 't' => [10, 20] }, annotation['parsed_fragment'])
-
-    query = %{
-      query {
-        project_media(ids: "#{pm.id},#{p.id}") {
-          clips: annotations(first: 10000, annotation_type: "clip") {
-            edges {
-              node {
-                ... on Dynamic {
-                  id
-                  data
-                  parsed_fragment
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-    post :create, params: { query: query, team: pm.team.slug }
-    assert_response :success
-    clips = JSON.parse(@response.body)['data']['project_media']['clips']['edges']
-    assert_equal 1, clips.size
-    assert_equal 'Clip Label', clips[0]['node']['data']['label']
-    assert_equal({ 't' => [10, 20] }, clips[0]['node']['parsed_fragment'])
-  end
-
-  test "should get team user from user" do
-    u = create_user
-    t = create_team
-    tu = create_team_user user: u, team: t
-    authenticate_with_user(u)
-
-    query = 'query { me { team_user(team_slug: "' + t.slug + '") { dbid } } }'
-    post :create, params: { query: query }
-    assert_response :success
-    assert_equal tu.id, JSON.parse(@response.body)['data']['me']['team_user']['dbid']
-
-    query = 'query { me { team_user(team_slug: "' + random_string + '") { dbid } } }'
-    post :create, params: { query: query }
-    assert_response :success
-    assert_nil JSON.parse(@response.body)['data']['me']['team_user']
-  end
-
-  test "should create report with image" do
-    create_report_design_annotation_type
-    u = create_user is_admin: true
-    pm = create_project_media
-    authenticate_with_user(u)
-    path = File.join(Rails.root, 'test', 'data', 'rails.png')
-    file = Rack::Test::UploadedFile.new(path, 'image/png')
-    query = 'mutation create { createDynamicAnnotationReportDesign(input: { action: "save", clientMutationId: "1", annotated_type: "ProjectMedia", annotated_id: "' + pm.id.to_s + '", set_fields: "{\"options\":{\"language\":\"en\"}}" }) { dynamic { dbid } } }'
-    post :create, params: { query: query, file: [file] }
-    assert_response :success
-    d = Dynamic.find(JSON.parse(@response.body)['data']['createDynamicAnnotationReportDesign']['dynamic']['dbid']).data.with_indifferent_access
-    assert_match /rails\.png/, d[:options]['image']
-  end
-
-  test "should define team languages settings" do
-    u = create_user is_admin: true
-    authenticate_with_user(u)
-    t = create_team
-    t.set_language nil
-    t.set_languages nil
-    t.save!
-
-    assert_nil t.reload.get_language
-    assert_nil t.reload.get_languages
-
-    query = 'mutation { updateTeam(input: { clientMutationId: "1", id: "' + t.graphql_id + '", language: "port" }) { team { id } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response 400
-
-    query = 'mutation { updateTeam(input: { clientMutationId: "1", id: "' + t.graphql_id + '", languages: "[\"port\"]" }) { team { id } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response 400
-
-    assert_nil t.reload.get_language
-    assert_nil t.reload.get_languages
-
-    query = 'mutation { updateTeam(input: { clientMutationId: "1", id: "' + t.graphql_id + '", language: "pt_BR", languages: "[\"es\", \"pt\", \"bho\"]" }) { team { id } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-
-    assert_equal 'pt_BR', t.reload.get_language
-    assert_equal ['es', 'pt', 'bho'], t.reload.get_languages
-  end
-
-  test "should define team custom statuses" do
-    u = create_user is_admin: true
-    authenticate_with_user(u)
-    t = create_team
-
-    custom_statuses = {
-      label: 'Field label',
-      active: '2',
-      default: '1',
-      statuses: [
-        { id: '1', locales: { en: { label: 'Custom Status 1', description: 'The meaning of this status' } }, style: { color: 'red' } },
-        { id: '2', locales: { en: { label: 'Custom Status 2', description: 'The meaning of that status' } }, style: { color: 'blue' } }
-      ]
-    }
-
-    query = 'mutation { updateTeam(input: { clientMutationId: "1", id: "' + t.graphql_id + '", language: "pt_BR", media_verification_statuses: ' + custom_statuses.to_json.to_json + ' }) { team { id, verification_statuses_with_counters: verification_statuses(items_count_for_status: "1", published_reports_count_for_status: "1"), verification_statuses } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    data = JSON.parse(@response.body).dig('data', 'updateTeam', 'team')
-    assert_match /items_count/, data['verification_statuses_with_counters'].to_json
-    assert_no_match /items_count:0/, data['verification_statuses'].to_json
-  end
-
-  test "should filter by user in ElasticSearch" do
-    u = create_user
-    t = create_team
-    create_team_user user: u, team: t
-    pm = create_project_media team: t, quote: 'This is a test', media: nil, user: u, disable_es_callbacks: false
-    create_project_media team: t, user: u, disable_es_callbacks: false
-    create_project_media team: t, disable_es_callbacks: false
-    sleep 1
-    authenticate_with_user(u)
-
-    query = 'query CheckSearch { search(query: "{\"keyword\":\"test\",\"users\":[' + u.id.to_s + ']}") { medias(first: 10) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug }
-
-    assert_response :success
-    assert_equal [pm.id], JSON.parse(@response.body)['data']['search']['medias']['edges'].collect{ |x| x['node']['dbid'] }
-  end
-
-  test "should filter by user in PostgreSQL" do
-    u = create_user
-    t = create_team
-    create_team_user user: u, team: t
-    p = create_project team: t
-    pm = create_project_media team: t, project: p, user: u
-    create_project_media team: t
-    authenticate_with_user(u)
-
-    query = 'query CheckSearch { search(query: "{\"users\":[' + u.id.to_s + '], \"projects\":[' + p.id.to_s + ']}") { medias(first: 10) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug }
-
-    assert_response :success
-    assert_equal [pm.id], JSON.parse(@response.body)['data']['search']['medias']['edges'].collect{ |x| x['node']['dbid'] }
-  end
-
-  test "should get project using API key" do
-    t = create_team
-    a = create_api_key
-    b = create_bot_user api_key_id: a.id, team: t
-    p = create_project team: t
-    authenticate_with_token(a)
-
-    query = 'query { me { dbid } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal b.id, JSON.parse(@response.body)['data']['me']['dbid']
-
-    query = 'query { project(id: "' + p.id.to_s + '") { dbid } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal p.id, JSON.parse(@response.body)['data']['project']['dbid']
-  end
-
-  test "should mark item as read" do
-    pm = create_project_media
-    assert !pm.reload.read
-    u = create_user is_admin: true
-    authenticate_with_user(u)
-
-    assert_difference 'ProjectMediaUser.count' do
-      query = 'mutation { createProjectMediaUser(input: { clientMutationId: "1", project_media_id: ' + pm.id.to_s + ', read: true }) { project_media { is_read } } }'
-      post :create, params: { query: query, team: pm.team.slug }
-      assert_response :success
-      assert pm.reload.read
-
-      query = 'mutation { createProjectMediaUser(input: { clientMutationId: "1", project_media_id: ' + pm.id.to_s + ', read: true }) { project_media { is_read } } }'
-      post :create, params: { query: query, team: pm.team.slug }
-      assert_response 409
-    end
-  end
-
-  test "should return if item is read" do
-    u = create_user is_admin: true
-    t = create_team
-    p = create_project team: t
-    pm1 = create_project_media project: p
-    ProjectMediaUser.create! user: u, project_media: pm1, read: true
-    pm2 = create_project_media project: p
-    ProjectMediaUser.create! user: create_user, project_media: pm2, read: true
-    pm3 = create_project_media project: p
-    authenticate_with_user(u)
-
-    {
-      pm1.id => [true, true],
-      pm2.id => [true, false],
-      pm3.id => [false, false]
-    }.each do |id, values|
-      ids = [id, p.id, t.id].join(',')
-      query = 'query { project_media(ids: "' + ids + '") { read_by_someone: is_read, read_by_me: is_read(by_me: true) } }'
-      post :create, params: { query: query, team: t.slug }
-      assert_response :success
-      data = JSON.parse(@response.body)['data']['project_media']
-      assert_equal values[0], data['read_by_someone']
-      assert_equal values[1], data['read_by_me']
-    end
-  end
-
-  test "should filter by read in ElasticSearch" do
-    u = create_user
-    t = create_team
-    create_team_user user: u, team: t
-    pm1 = create_project_media team: t, quote: 'This is a test', media: nil, read: true, disable_es_callbacks: false
-    pm2 = create_project_media team: t, quote: 'This is another test', media: nil, disable_es_callbacks: false
-    pm3 = create_project_media quote: 'This is another test', media: nil, disable_es_callbacks: false
-    sleep 1
-    authenticate_with_user(u)
-
-    query = 'query CheckSearch { search(query: "{\"keyword\":\"test\",\"read\":[1]}") { medias(first: 10) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal [pm1.id], JSON.parse(@response.body)['data']['search']['medias']['edges'].collect{ |x| x['node']['dbid'] }
-
-    query = 'query CheckSearch { search(query: "{\"keyword\":\"test\",\"read\":[0]}") { medias(first: 10) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal [pm2.id], JSON.parse(@response.body)['data']['search']['medias']['edges'].collect{ |x| x['node']['dbid'] }
-
-    query = 'query CheckSearch { search(query: "{\"keyword\":\"test\"}") { medias(first: 10) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal [pm1.id, pm2.id].sort, JSON.parse(@response.body)['data']['search']['medias']['edges'].collect{ |x| x['node']['dbid'] }.sort
-  end
-
-  test "should filter by read in PostgreSQL" do
-    u = create_user
-    t = create_team
-    create_team_user user: u, team: t
-    p = create_project team: t
-    pm1 = create_project_media team: t, project: p, read: true
-    pm2 = create_project_media team: t, project: p
-    pm3 = create_project_media
-    authenticate_with_user(u)
-
-    query = 'query CheckSearch { search(query: "{\"read\":true, \"projects\":[' + p.id.to_s + ']}") { medias(first: 10) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal [pm1.id], JSON.parse(@response.body)['data']['search']['medias']['edges'].collect{ |x| x['node']['dbid'] }
-
-    query = 'query CheckSearch { search(query: "{\"read\":false, \"projects\":[' + p.id.to_s + ']}") { medias(first: 10) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal [pm2.id], JSON.parse(@response.body)['data']['search']['medias']['edges'].collect{ |x| x['node']['dbid'] }
-
-    query = 'query CheckSearch { search(query: "{\"projects\":[' + p.id.to_s + ']}") { medias(first: 10) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal [pm1.id, pm2.id].sort, JSON.parse(@response.body)['data']['search']['medias']['edges'].collect{ |x| x['node']['dbid'] }.sort
-  end
-
-  test "should get team fieldsets" do
-    u = create_user is_admin: true
-    authenticate_with_user(u)
-    t = create_team
-    query = 'query { team { get_fieldsets } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_kind_of Array, JSON.parse(@response.body)['data']['team']['get_fieldsets']
-  end
-
-  test "should get item tasks by fieldset" do
-    u = create_user is_admin: true
-    t = create_team
-    pm = create_project_media team: t
-    t1 = create_task annotated: pm, fieldset: 'tasks'
-    t2 = create_task annotated: pm, fieldset: 'metadata'
-    ids = [pm.id, nil, t.id].join(',')
-    authenticate_with_user(u)
-
-    query = 'query { project_media(ids: "' + ids + '") { tasks(fieldset: "tasks", first: 1000) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal t1.id, JSON.parse(@response.body)['data']['project_media']['tasks']['edges'][0]['node']['dbid'].to_i
-
-    query = 'query { project_media(ids: "' + ids + '") { tasks(fieldset: "metadata", first: 1000) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal t2.id, JSON.parse(@response.body)['data']['project_media']['tasks']['edges'][0]['node']['dbid'].to_i
-    s = create_source team: t
-    t3 = create_task annotated: s, fieldset: 'metadata'
-    query = 'query { source(id: "' + s.id.to_s + '") { tasks(fieldset: "metadata", first: 1000) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal t3.id, JSON.parse(@response.body)['data']['source']['tasks']['edges'][0]['node']['dbid'].to_i
-  end
-
-  test "should get team tasks by fieldset" do
-    u = create_user is_admin: true
-    t = create_team
-    t1 = create_team_task team_id: t.id, fieldset: 'tasks'
-    t2 = create_team_task team_id: t.id, fieldset: 'metadata'
-    authenticate_with_user(u)
-
-    query = 'query { team { team_tasks(fieldset: "tasks", first: 1000) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal t1.id, JSON.parse(@response.body)['data']['team']['team_tasks']['edges'][0]['node']['dbid'].to_i
-
-    query = 'query { team { team_tasks(fieldset: "metadata", first: 1000) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal t2.id, JSON.parse(@response.body)['data']['team']['team_tasks']['edges'][0]['node']['dbid'].to_i
-  end
-
-  test "should update task options" do
-    u = create_user
-    t = create_team
-    create_team_user team: t, user: u, role: 'admin'
-    tt = create_team_task team_id: t.id, label: 'Select one', type: 'single_choice', options: ['ans_a', 'ans_b', 'ans_c']
-    pm = create_project_media team: t
-    authenticate_with_user(u)
-    query = 'mutation { updateTeamTask(input: { clientMutationId: "1", id: "' + tt.graphql_id + '", label: "Select only one", json_options: "[ \"bli\", \"blo\", \"bla\" ]", options_diff: "{ \"deleted\": [\"ans_c\"], \"changed\": { \"ans_a\": \"bli\", \"ans_b\": \"blo\" }, \"added\": \"bla\" }" }) { team_task { label } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal 'Select only one', tt.reload.label
-    assert_equal ['bli', 'blo', 'bla'], tt.options
-  end
-
-  test "should replace blank project media by another" do
-    u = create_user
-    t = create_team
-    create_team_user team: t, user: u, role: 'admin'
-    old = create_project_media team: t, media: Blank.create!
-    r = publish_report(old)
-    new = create_project_media team: t
-    authenticate_with_user(u)
-
-    query = 'mutation { replaceProjectMedia(input: { clientMutationId: "1", project_media_to_be_replaced_id: "' + old.graphql_id + '", new_project_media_id: "' + new.graphql_id + '" }) { old_project_media_deleted_id, new_project_media { dbid } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    data = JSON.parse(@response.body)['data']['replaceProjectMedia']
-    assert_equal old.graphql_id, data['old_project_media_deleted_id']
-    assert_equal new.id, data['new_project_media']['dbid']
-    assert_nil ProjectMedia.find_by_id(old.id)
-    assert_equal r, new.get_dynamic_annotation('report_design')
-  end
-
-  test "should set and get Slack settings for team" do
-    u = create_user
-    t = create_team
-    create_team_user team: t, user: u, role: 'admin'
-    authenticate_with_user(u)
-    query = 'mutation { updateTeam(input: { clientMutationId: "1", id: "' + t.graphql_id + '", slack_notifications: "[{\"label\":\"not #1\",\"event_type\":\"any_activity\",\"slack_channel\":\"#list\"}]" }) { team { get_slack_notifications } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_not_nil JSON.parse(@response.body)['data']['updateTeam']['team']['get_slack_notifications']
-  end
-
-  test "should set and get special list filters for team" do
-    u = create_user
-    t = create_team
-    create_team_user team: t, user: u, role: 'admin'
-    authenticate_with_user(u)
-    # Tipline list
-    query = 'mutation { updateTeam(input: { clientMutationId: "1", id: "' + t.graphql_id + '", tipline_inbox_filters: "{\"read\":[\"0\"],\"projects\":[\"-1\"]}" }) { team { get_tipline_inbox_filters } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_not_nil JSON.parse(@response.body)['data']['updateTeam']['team']['get_tipline_inbox_filters']
-    # Suggested match list
-    query = 'mutation { updateTeam(input: { clientMutationId: "1", id: "' + t.graphql_id + '", suggested_matches_filters: "{\"projects\":[\"-1\"],\"suggestions_count\":{\"min\":5,\"max\":10}}" }) { team { get_suggested_matches_filters } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_not_nil JSON.parse(@response.body)['data']['updateTeam']['team']['get_suggested_matches_filters']
-  end
-
-  test "should get Smooch Bot RSS feed preview if has permissions" do
-    rss = %{
-      <rss xmlns:atom="http://www.w3.org/2005/Atom" version="2.0">
-        <channel>
-          <title>Test</title>
-          <link>http://test.com/rss.xml</link>
-          <description>Test</description>
-          <language>en</language>
-          <lastBuildDate>Fri, 09 Oct 2020 18:00:48 GMT</lastBuildDate>
-          <managingEditor>test@test.com (editors)</managingEditor>
-          <item>
-            <title>Foo</title>
-            <description>This is the description.</description>
-            <pubDate>Wed, 11 Apr 2018 15:25:00 GMT</pubDate>
-            <link>http://foo</link>
-          </item>
-          <item>
-            <title>Bar</title>
-            <description>This is the description.</description>
-            <pubDate>Wed, 10 Apr 2018 15:25:00 GMT</pubDate>
-            <link>http://bar</link>
-          </item>
-        </channel>
-      </rss>
-    }
-    u = create_user
-    t = create_team
-    b = create_team_bot name: 'Smooch', login: 'smooch', set_approved: true, set_events: [], set_request_url: "#{CheckConfig.get('checkdesk_base_url_private')}/api/bots/smooch"
-    tbi = create_team_bot_installation team_id: t.id, user_id: b.id
-    tu = create_team_user team: t, user: u, role: 'admin'
-    authenticate_with_user(u)
-    url = random_url
-    WebMock.stub_request(:get, url).to_return(status: 200, body: rss)
-    output = "Foo\nhttp://foo\n\nBar\nhttp://bar"
-    query = 'query { node(id: "' + tbi.graphql_id + '") { ... on TeamBotInstallation { smooch_bot_preview_rss_feed(rss_feed_url: "' + url + '", number_of_articles: 3) } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_no_match /Sorry/, @response.body
-    assert_equal output, JSON.parse(@response.body)['data']['node']['smooch_bot_preview_rss_feed']
-  end
-
-  test "should set and get Alegre Bot settings" do
-    u = create_user
-    t = create_team
-    b = create_team_bot name: 'Alegre', login: 'alegre', set_approved: true
-    tbi = create_team_bot_installation team_id: t.id, user_id: b.id
-    tu = create_team_user team: t, user: u, role: 'admin'
-    authenticate_with_user(u)
-    query = 'mutation { updateTeamBotInstallation(input: { clientMutationId: "1", id: "' + tbi.graphql_id + '", json_settings: "{\"text_length_matching_threshold\":\"4\"}" }) { team_bot_installation { json_settings } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    query = 'query { node(id: "' + tbi.graphql_id + '") { ... on TeamBotInstallation { alegre_settings } } }'
-    post :create, params: { query: query, team: t.slug }
-    alegre_settings =  JSON.parse(@response.body)['data']['node']['alegre_settings']
-    assert_equal 4.0, alegre_settings['text_length_matching_threshold']
-  end
-
-  test "should not get Smooch Bot RSS feed preview if not owner" do
-    u = create_user
-    t = create_team
-    b = create_team_bot name: 'Smooch', login: 'smooch', set_approved: true, set_events: [], set_request_url: "#{CheckConfig.get('checkdesk_base_url_private')}/api/bots/smooch"
-    tbi = create_team_bot_installation team_id: t.id, user_id: b.id
-    tu = create_team_user team: t, user: u, role: 'collaborator'
-    authenticate_with_user(u)
-    url = random_url
-    output = "Foo\nhttp://foo\n\nBar\nhttp://bar"
-    query = 'query { node(id: "' + tbi.graphql_id + '") { ... on TeamBotInstallation { smooch_bot_preview_rss_feed(rss_feed_url: "' + url + '", number_of_articles: 3) } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_match /Sorry/, @response.body
-  end
-
-  test "should not get Smooch Bot RSS feed preview if not member of the team" do
-    u = create_user
-    t = create_team
-    b = create_team_bot name: 'Smooch', login: 'smooch', set_approved: true, set_events: [], set_request_url: "#{CheckConfig.get('checkdesk_base_url_private')}/api/bots/smooch"
-    tbi = create_team_bot_installation team_id: t.id, user_id: b.id
-    tu = create_team_user team: t, user: u, role: 'admin'
-    authenticate_with_user(create_user)
-    url = random_url
-    output = "Foo\nhttp://foo\n\nBar\nhttp://bar"
-    query = 'query { node(id: "' + tbi.graphql_id + '") { ... on TeamBotInstallation { smooch_bot_preview_rss_feed(rss_feed_url: "' + url + '", number_of_articles: 3) } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_match /Sorry/, @response.body
-  end
-
-  test "should not get Smooch Bot RSS feed preview if not logged in" do
-    u = create_user
-    t = create_team
-    b = create_team_bot name: 'Smooch', login: 'smooch', set_approved: true, set_events: [], set_request_url: "#{CheckConfig.get('checkdesk_base_url_private')}/api/bots/smooch"
-    tbi = create_team_bot_installation team_id: t.id, user_id: b.id
-    tu = create_team_user team: t, user: u, role: 'admin'
-    url = random_url
-    output = "Foo\nhttp://foo\n\nBar\nhttp://bar"
-    query = 'query { node(id: "' + tbi.graphql_id + '") { ... on TeamBotInstallation { smooch_bot_preview_rss_feed(rss_feed_url: "' + url + '", number_of_articles: 3) } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_match /Sorry/, @response.body
-  end
-
-  test "should return similar items" do
-    u = create_user
-    t = create_team
-    create_team_user team: t, user: u, role: 'admin'
-    authenticate_with_user(u)
-    p = create_project team: t
-    p1 = create_project_media project: p
-    p1a = create_project_media project: p
-    p1b = create_project_media project: p
-    create_relationship source_id: p1.id, target_id: p1a.id, relationship_type: Relationship.suggested_type
-    create_relationship source_id: p1.id, target_id: p1b.id, relationship_type: Relationship.suggested_type
-    p2 = create_project_media project: p
-    p2a = create_project_media project: p
-    p2b = create_project_media project: p
-    create_relationship source_id: p2.id, target_id: p2a.id
-    create_relationship source_id: p2.id, target_id: p2b.id, relationship_type: Relationship.suggested_type
-    post :create, params: { query: "query { project_media(ids: \"#{p1.id},#{p.id}\") { is_main, is_secondary, is_confirmed_similar_to_another_item, suggested_main_item { id }, confirmed_main_item { id }, default_relationships_count, default_relationships(first: 10000) { edges { node { dbid } } }, confirmed_similar_relationships(first: 10000) { edges { node { dbid } } }, suggested_similar_relationships(first: 10000) { edges { node { target { dbid } } } } } }", team: t.slug }
-    assert_equal [p1a.id, p1b.id].sort, JSON.parse(@response.body)['data']['project_media']['suggested_similar_relationships']['edges'].collect{ |x| x['node']['target']['dbid'] }.sort
-  end
-
-  test "should return suggested similar items count" do
-    u = create_user
-    t = create_team
-    create_team_user team: t, user: u, role: 'admin'
-    authenticate_with_user(u)
-    p = create_project team: t
-    p1 = create_project_media project: p
-    p1a = create_project_media project: p
-    p1b = create_project_media project: p
-    create_relationship source_id: p1.id, target_id: p1a.id, relationship_type: Relationship.suggested_type
-    create_relationship source_id: p1.id, target_id: p1b.id, relationship_type: Relationship.suggested_type
-    p2 = create_project_media project: p
-    p2a = create_project_media project: p
-    p2b = create_project_media project: p
-    create_relationship source_id: p2.id, target_id: p2a.id
-    create_relationship source_id: p2.id, target_id: p2b.id, relationship_type: Relationship.suggested_type
-    post :create, params: { query: "query { project_media(ids: \"#{p1.id},#{p.id}\") { suggested_similar_items_count } }", team: t.slug }; false
-    assert_equal 2, JSON.parse(@response.body)['data']['project_media']['suggested_similar_items_count']
-  end
-
-  test "should return confirmed similar items count" do
-    u = create_user
-    t = create_team
-    create_team_user team: t, user: u, role: 'admin'
-    authenticate_with_user(u)
-    p = create_project team: t
-    p1 = create_project_media project: p
-    p1a = create_project_media project: p
-    p1b = create_project_media project: p
-    create_relationship source_id: p1.id, target_id: p1a.id, relationship_type: Relationship.confirmed_type
-    create_relationship source_id: p1.id, target_id: p1b.id, relationship_type: Relationship.confirmed_type
-    p2 = create_project_media project: p
-    p2a = create_project_media project: p
-    p2b = create_project_media project: p
-    create_relationship source_id: p2.id, target_id: p2a.id
-    create_relationship source_id: p2.id, target_id: p2b.id, relationship_type: Relationship.confirmed_type
-    post :create, params: { query: "query { project_media(ids: \"#{p1.id},#{p.id}\") { confirmed_similar_items_count } }", team: t.slug }; false
-    assert_equal 2, JSON.parse(@response.body)['data']['project_media']['confirmed_similar_items_count']
-  end
-
-  test "should sort search by metadata value where items without metadata value show first on ascending order" do
-    RequestStore.store[:skip_cached_field_update] = false
-    at = create_annotation_type annotation_type: 'task_response_free_text'
-    create_field_instance annotation_type_object: at, name: 'response_free_text'
-    u = create_user is_admin: true
-    t = create_team
-    create_team_user team: t, user: u
-    tt1 = create_team_task fieldset: 'metadata', team_id: t.id
-    tt2 = create_team_task fieldset: 'metadata', team_id: t.id
-    t.list_columns = ["task_value_#{tt1.id}", "task_value_#{tt2.id}"]
-    t.save!
-    pm1 = create_project_media team: t, disable_es_callbacks: false
-    pm2 = create_project_media team: t, disable_es_callbacks: false
-    pm3 = create_project_media team: t, disable_es_callbacks: false
-    pm4 = create_project_media team: t, disable_es_callbacks: false
-
-    m = pm1.get_annotations('task').map(&:load).select{ |t| t.team_task_id == tt1.id }.last
-    m.disable_es_callbacks = false
-    m.response = { annotation_type: 'task_response_free_text', set_fields: { response_free_text: 'B' }.to_json }.to_json
-    m.save!
-    m = pm3.get_annotations('task').map(&:load).select{ |t| t.team_task_id == tt1.id }.last
-    m.disable_es_callbacks = false
-    m.response = { annotation_type: 'task_response_free_text', set_fields: { response_free_text: 'A' }.to_json }.to_json
-    m.save!
-    m = pm4.get_annotations('task').map(&:load).select{ |t| t.team_task_id == tt1.id }.last
-    m.disable_es_callbacks = false
-    m.response = { annotation_type: 'task_response_free_text', set_fields: { response_free_text: 'C' }.to_json }.to_json
-    m.save!
-    sleep 5
-
-    m = pm1.get_annotations('task').map(&:load).select{ |t| t.team_task_id == tt2.id }.last
-    m.disable_es_callbacks = false
-    m.response = { annotation_type: 'task_response_free_text', set_fields: { response_free_text: 'C' }.to_json }.to_json
-    m.save!
-    m = pm2.get_annotations('task').map(&:load).select{ |t| t.team_task_id == tt2.id }.last
-    m.disable_es_callbacks = false
-    m.response = { annotation_type: 'task_response_free_text', set_fields: { response_free_text: 'B' }.to_json }.to_json
-    m.save!
-    m = pm4.get_annotations('task').map(&:load).select{ |t| t.team_task_id == tt2.id }.last
-    m.disable_es_callbacks = false
-    m.response = { annotation_type: 'task_response_free_text', set_fields: { response_free_text: 'A' }.to_json }.to_json
-    m.save!
-    sleep 5
-
-    authenticate_with_user(u)
-
-    query = 'query CheckSearch { search(query: "{\"sort\":\"task_value_' + tt1.id.to_s + '\",\"sort_type\":\"asc\"}") {medias(first:20){edges{node{dbid}}}}}'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    results = JSON.parse(@response.body)['data']['search']['medias']['edges'].collect{ |x| x['node']['dbid'] }
-    assert_equal 4, results.size
-    assert_equal pm2.id, results.first
-
-    query = 'query CheckSearch { search(query: "{\"sort\":\"task_value_' + tt2.id.to_s + '\",\"sort_type\":\"asc\"}") {medias(first:20){edges{node{dbid}}}}}'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    results = JSON.parse(@response.body)['data']['search']['medias']['edges'].collect{ |x| x['node']['dbid'] }
-    assert_equal 4, results.size
-    assert_equal pm3.id, results.first
-
-    query = 'query CheckSearch { search(query: "{\"sort\":\"task_value_' + tt1.id.to_s + '\",\"sort_type\":\"desc\"}") {medias(first:20){edges{node{dbid}}}}}'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    results = JSON.parse(@response.body)['data']['search']['medias']['edges'].collect{ |x| x['node']['dbid'] }
-    assert_equal 4, results.size
-    assert_equal pm2.id, results.last
-
-    query = 'query CheckSearch { search(query: "{\"sort\":\"task_value_' + tt2.id.to_s + '\",\"sort_type\":\"desc\"}") {medias(first:20){edges{node{dbid}}}}}'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    results = JSON.parse(@response.body)['data']['search']['medias']['edges'].collect{ |x| x['node']['dbid'] }
-    assert_equal 4, results.size
-    assert_equal pm3.id, results.last
-  end
-
-  test "should load permissions for GraphQL" do
-    pm1 = create_project_media
-    pm2 = create_project_media
-    User.current = create_user
-    PermissionsLoader.any_instance.stubs(:fulfill).returns(nil)
-    assert_kind_of Array, PermissionsLoader.new(nil).perform([pm1.id, pm2.id])
-    PermissionsLoader.any_instance.unstub(:fulfill)
-    User.current = nil
-  end
 end
