+++ test/controllers/graphql_controller_4_test.rb
@@ -45,15 +45,6 @@ class GraphqlController4Test < ActionController::TestCase
     @pms.each { |pm| assert_equal CheckArchivedFlags::FlagCodes::NONE, pm.reload.archived }
   end
 
-  test "should not bulk-send project medias to trash if there are more than 10.000 ids" do
-    ids = []
-    10001.times { ids << random_string }
-    query = 'mutation { updateProjectMedias(input: { clientMutationId: "1", ids: ' + ids.to_json + ', action: "archived", params: "{\"archived:\": 1}" }) { ids, team { dbid } } }'
-    post :create, params: { query: query, team: @t.slug }
-    assert_response 400
-    assert_error_message 'maximum'
-  end
-
   test "should bulk-send project medias to trash" do
     @pms.each { |pm| assert_equal CheckArchivedFlags::FlagCodes::NONE, pm.archived }
     @ps.each { |p| assert_equal 1, p.reload.medias_count }
@@ -321,481 +312,6 @@ class GraphqlController4Test < ActionController::TestCase
     assert_equal p4.title, Rails.cache.read("check_cached_field:ProjectMedia:#{t_pm1.id}:folder")
     assert_equal p4.title, Rails.cache.read("check_cached_field:ProjectMedia:#{t2_pm1.id}:folder")
     assert_equal p4.title, Rails.cache.read("check_cached_field:ProjectMedia:#{t_pm2.id}:folder")
-  end
-
-  test "should update archived media by owner" do
-    pm = create_project_media team: @t, archived: CheckArchivedFlags::FlagCodes::TRASHED
-    query = "mutation { updateProjectMedia(input: { clientMutationId: \"1\", id: \"#{pm.graphql_id}\"}) { project_media { permissions } } }"
-    post :create, params: { query: query, team: @t.slug }
-    assert_response :success
-    data = JSON.parse(@response.body)['data']['updateProjectMedia']['project_media']
-    permissions = JSON.parse(data['permissions'])
-    assert_equal true, permissions['update ProjectMedia']
-  end
-
-  test "should update project media source" do
-    s = create_source team: @t
-    s2 = create_source team: @t
-    pm = create_project_media team: @t, source_id: s.id, skip_autocreate_source: false
-    pm2 = create_project_media team: @t, source_id: s2.id, skip_autocreate_source: false
-    assert_equal s.id, pm.source_id
-    query = "mutation { updateProjectMedia(input: { clientMutationId: \"1\", id: \"#{pm.graphql_id}\", source_id: #{s2.id}}) { project_media { source { dbid, medias_count, medias(first: 10) { edges { node { dbid } } } } } } }"
-    post :create, params: { query: query, team: @t.slug }
-    assert_response :success
-    data = JSON.parse(@response.body)['data']['updateProjectMedia']['project_media']
-    assert_equal s2.id, data['source']['dbid']
-    assert_equal 2, data['source']['medias_count']
-    assert_equal 2, data['source']['medias']['edges'].size
-  end
-
-  test "should set a project as default" do
-    default_folder = @t.default_folder.id
-    p = create_project team: @t
-    query = "mutation { updateProject(input: { clientMutationId: \"1\", id: \"#{p.graphql_id}\", previous_default_project_id: #{default_folder},is_default: true}) { project { is_default }, previous_default_project { dbid, is_default } } }"
-    post :create, params: { query: query, team: @t.slug }
-    assert_response :success
-    data = JSON.parse(@response.body)['data']['updateProject']
-    assert data['project']['is_default']
-    assert_equal default_folder, data['previous_default_project']['dbid']
-    assert !data['previous_default_project']['is_default']
-  end
-
-  test "should create related project media for source" do
-    t = create_team
-    pm = create_project_media team: t
-    u = create_user
-    create_team_user user: u, team: t, role: 'admin'
-    authenticate_with_user(u)
-    query = 'mutation create { createSource(input: { name: "new source", slogan: "new source", clientMutationId: "1", add_to_project_media_id: ' + pm.id.to_s + ' }) { source { dbid } } }'
-    post :create, params: { query: query, team: t }
-    assert_response :success
-    source = JSON.parse(@response.body)['data']['createSource']['source']
-    assert_equal pm.reload.source_id, source['dbid']
-  end
-
-  test "should search team sources by keyword" do
-    t = create_team slug: 'sawy'
-    u = create_user
-    create_team_user team: t, user: u, role: 'admin'
-    create_source team: t, name: 'keyword begining'
-    create_source team: t, name: 'ending keyword'
-    create_source team: t, name: 'in the KEYWORD middle'
-    create_source team: t
-    authenticate_with_user(u)
-    query = 'query read { team(slug: "sawy") { sources_count, sources(first: 1000) { edges { node { dbid } } } } }'
-    post :create, params: { query: query }
-    assert_response :success
-    edges = JSON.parse(@response.body)['data']['team']['sources']['edges']
-    assert_equal 4, edges.length
-    query = 'query read { team(slug: "sawy") { sources_count(keyword: "keyword"), sources(first: 1000, keyword: "keyword") { edges { node { dbid } } } } }'
-    post :create, params: { query: query }
-    assert_response :success
-    edges = JSON.parse(@response.body)['data']['team']['sources']['edges']
-    assert_equal 3, edges.length
-  end
-
-  test "should update last_active_at from users before a graphql request" do
-    assert_nil @u.last_active_at
-    query = "query { user(id: #{@u.id}) { last_active_at } }"
-    post :create, params: { query: query }
-    assert_response :success
-    assert_not_nil JSON.parse(@response.body)['data']['user']['last_active_at']
-    assert_not_nil @u.reload.last_active_at
-  end
-
-  test "should not get Smooch integrations if not permissioned" do
-    t = create_team private: true
-    b = create_team_bot login: 'smooch', set_approved: true
-    app_id = random_string
-    tbi = create_team_bot_installation team_id: t.id, user_id: b.id, settings: { smooch_app_id: app_id }
-    u = create_user
-    t2 = create_team
-    create_team_user user: u, team: t2, role: 'admin'
-    
-    query = "query { team(slug: \"#{t.slug}\") { team_bot_installations(first: 1) { edges { node { smooch_enabled_integrations } } } } }"
-    post :create, params: { query: query }
-    assert_error_message 'Not Found'
-
-    authenticate_with_user(u)
-    post :create, params: { query: query }
-    assert_error_message 'Not Found'
-  end
-
-  test "should get Smooch integrations if permissioned" do
-    t = create_team private: true
-    b = create_team_bot login: 'smooch', set_approved: true
-    app_id = random_string
-    tbi = create_team_bot_installation team_id: t.id, user_id: b.id, settings: { smooch_app_id: app_id }
-    u = create_user
-    create_team_user user: u, team: t, role: 'admin'
-    
-    authenticate_with_user(u)
-    query = "query { team(slug: \"#{t.slug}\") { team_bot_installations(first: 1) { edges { node { smooch_enabled_integrations } } } } }"
-    post :create, params: { query: query }
-    assert_not_nil json_response.dig('data', 'team', 'team_bot_installations', 'edges', 0, 'node', 'smooch_enabled_integrations')
-  end
-
-  test "should remove Smooch integration if permissioned" do
-    t = create_team private: true
-    b = create_team_bot login: 'smooch', set_approved: true
-    app_id = random_string
-    tbi = create_team_bot_installation team_id: t.id, user_id: b.id, settings: { smooch_app_id: app_id }
-    u = create_user
-    create_team_user user: u, team: t, role: 'admin'
-    
-    authenticate_with_user(u)
-    query = "mutation { smoochBotRemoveIntegration(input: { clientMutationId: \"1\", team_bot_installation_id: \"#{tbi.graphql_id}\", integration_type: \"whatsapp\" }) { team_bot_installation { smooch_enabled_integrations } } }"
-    post :create, params: { query: query }
-    assert_not_nil json_response.dig('data', 'smoochBotRemoveIntegration', 'team_bot_installation', 'smooch_enabled_integrations')
-  end
-
-  test "should not remove Smooch integration if not permissioned" do
-    t = create_team private: true
-    b = create_team_bot login: 'smooch', set_approved: true
-    app_id = random_string
-    tbi = create_team_bot_installation team_id: t.id, user_id: b.id, settings: { smooch_app_id: app_id }
-    u = create_user
-    t2 = create_team
-    create_team_user user: u, team: t2, role: 'admin'
-    query = "mutation { smoochBotRemoveIntegration(input: { clientMutationId: \"1\", team_bot_installation_id: \"#{tbi.graphql_id}\", integration_type: \"whatsapp\" }) { team_bot_installation { smooch_enabled_integrations } } }"
-    
-    post :create, params: { query: query }
-    assert_error_message 'Not Found'
-
-    authenticate_with_user(u)
-    post :create, params: { query: query }
-    assert_error_message 'Not Found'
-  end
-
-  test "should add Smooch integration if permissioned" do
-    SmoochApi::IntegrationApi.any_instance.stubs(:create_integration).returns(nil)
-    t = create_team private: true
-    b = create_team_bot login: 'smooch', set_approved: true
-    app_id = random_string
-    tbi = create_team_bot_installation team_id: t.id, user_id: b.id, settings: { smooch_app_id: app_id }
-    u = create_user
-    create_team_user user: u, team: t, role: 'admin'
-    
-    authenticate_with_user(u)
-    query = 'mutation { smoochBotAddIntegration(input: { clientMutationId: "1", team_bot_installation_id: "' + tbi.graphql_id + '", integration_type: "messenger", params: "{\"token\":\"abc\"}" }) { team_bot_installation { smooch_enabled_integrations } } }'
-    post :create, params: { query: query }
-    assert_not_nil json_response.dig('data', 'smoochBotAddIntegration', 'team_bot_installation', 'smooch_enabled_integrations')
-  end
-
-  test "should not add Smooch integration if not permissioned" do
-    t = create_team private: true
-    b = create_team_bot login: 'smooch', set_approved: true
-    app_id = random_string
-    tbi = create_team_bot_installation team_id: t.id, user_id: b.id, settings: { smooch_app_id: app_id }
-    u = create_user
-    t2 = create_team
-    create_team_user user: u, team: t2, role: 'admin'
-    query = 'mutation { smoochBotAddIntegration(input: { clientMutationId: "1", team_bot_installation_id: "' + tbi.graphql_id + '", integration_type: "messenger", params: "{\"token\":\"abc\"}" }) { team_bot_installation { smooch_enabled_integrations } } }'
-
-    post :create, params: { query: query }
-    assert_error_message 'Not Found'
-
-    authenticate_with_user(u)
-    post :create, params: { query: query }
-    assert_error_message 'Not Found'
-  end
-
-  test "should get saved search filters" do
-    t = create_team
-    ss = create_saved_search team: t, filters: { foo: 'bar' }
-    query = "query { team(slug: \"#{t.slug}\") { saved_searches(first: 1) { edges { node { filters } } } } }"
-    post :create, params: { query: query }
-    assert_equal '{"foo":"bar"}', JSON.parse(@response.body).dig('data', 'team', 'saved_searches', 'edges', 0, 'node', 'filters')
-    assert_response :success
-  end
-
-  test "should search by report status" do
-    setup_elasticsearch
-    RequestStore.store[:skip_cached_field_update] = false
-    t = create_team
-    u = create_user
-    create_team_user team: t, user: u, role: 'admin'
-    authenticate_with_user(u)
-
-    # Published
-    pm1 = create_project_media team: t, disable_es_callbacks: false
-    r1 = publish_report(pm1)
-    r1 = Dynamic.find(r1.id)
-    r1.disable_es_callbacks = false
-    r1.set_fields = { state: 'published' }.to_json
-    r1.save!
-
-    # Paused
-    pm2 = create_project_media team: t, disable_es_callbacks: false
-    r2 = publish_report(pm2)
-    r2 = Dynamic.find(r2.id)
-    r2.disable_es_callbacks = false
-    r2.set_fields = { state: 'paused' }.to_json
-    r2.save!
-
-    # Not published
-    pm3 = create_project_media team: t, disable_es_callbacks: false
-
-    # Search
-    sleep 2
-    query = 'query CheckSearch { search(query: "{\"report_status\":[\"published\"]}") { medias(first: 20) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal [pm1.id], JSON.parse(@response.body)['data']['search']['medias']['edges'].collect{ |e| e['node']['dbid'] }
-
-    # Verify filter by report published data
-    query = 'query CheckSearch { search(query: "{\"range\": {\"report_published_at\":{\"condition\":\"less_than\",\"period\":\"1\",\"period_type\":\"y\"},\"timezone\":\"America/Bahia\"}}") { medias(first:20){edges{node{dbid}}}}}'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    results = JSON.parse(@response.body)['data']['search']['medias']['edges'].collect{ |x| x['node']['dbid'] }
-    assert_equal [pm1.id, pm2.id], results.sort
-  end
-
-  test "should get a single bot installation" do
-    t = create_team private: true
-    b = create_team_bot login: 'smooch', set_approved: true
-    app_id = random_string
-    tbi = create_team_bot_installation team_id: t.id, user_id: b.id, settings: { smooch_app_id: app_id }
-    u = create_user
-    create_team_user user: u, team: t, role: 'admin'
-    
-    authenticate_with_user(u)
-    query = "query { team(slug: \"#{t.slug}\") { team_bot_installation(bot_identifier: \"smooch\") { smooch_enabled_integrations(force: true) } } }"
-    post :create, params: { query: query }
-    assert_not_nil json_response.dig('data', 'team', 'team_bot_installation', 'smooch_enabled_integrations')
-  end
-
-  test "should search using OR or AND on PG" do
-    t = create_team
-    p1 = create_project team: t
-    p2 = create_project team: t
-    u = create_user
-    create_team_user team: t, user: u, role: 'admin'
-    authenticate_with_user(u)
-
-    pm1 = create_project_media team: t, project: p1, read: true
-    pm2 = create_project_media team: t, project: p2, read: false
-
-    query = 'query CheckSearch { search(query: "{\"operator\":\"AND\",\"read\":true,\"projects\":[' + p2.id.to_s + ']}") { medias(first: 20) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal [], JSON.parse(@response.body)['data']['search']['medias']['edges'].collect{ |e| e['node']['dbid'] }
-
-    query = 'query CheckSearch { search(query: "{\"operator\":\"OR\",\"read\":true,\"projects\":[' + p2.id.to_s + ']}") { medias(first: 20) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal [pm1.id, pm2.id].sort, JSON.parse(@response.body)['data']['search']['medias']['edges'].collect{ |e| e['node']['dbid'] }.sort
-  end
-
-  test "should search using OR or AND on ES" do
-    setup_elasticsearch
-    RequestStore.store[:skip_cached_field_update] = false
-    t = create_team
-    p1 = create_project team: t
-    p2 = create_project team: t
-    u = create_user
-    create_team_user team: t, user: u, role: 'admin'
-    authenticate_with_user(u)
-
-    pm1 = create_project_media team: t, project: p1, read: true
-    pm2 = create_project_media team: t, project: p2, read: false
-
-    query = 'query CheckSearch { search(query: "{\"operator\":\"AND\",\"read\":[1],\"projects\":[' + p2.id.to_s + '],\"report_status\":\"unpublished\"}") { medias(first: 20) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal [], JSON.parse(@response.body)['data']['search']['medias']['edges'].collect{ |e| e['node']['dbid'] }
-
-    query = 'query CheckSearch { search(query: "{\"operator\":\"OR\",\"read\":[1],\"projects\":[' + p2.id.to_s + '],\"report_status\":\"unpublished\"}") { medias(first: 20) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal [pm1.id, pm2.id].sort, JSON.parse(@response.body)['data']['search']['medias']['edges'].collect{ |e| e['node']['dbid'] }.sort
-  end
-
-  test "should search by project" do
-    t = create_team
-    p = create_project team: t
-    u = create_user
-    create_team_user team: t, user: u, role: 'admin'
-    authenticate_with_user(u)
-
-    create_project_media team: t, project: nil, project_id: nil
-    create_project_media project: p
-    
-    query = 'query CheckSearch { search(query: "{}") { medias(first: 20) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug }
-    assert_response :success
-    assert_equal 2, JSON.parse(@response.body)['data']['search']['medias']['edges'].size
-  end
-
-  test "should get Smooch newsletter information" do
-    setup_smooch_bot(true)
-    rss = '<rss version="1"><channel><title>x</title><link>x</link><description>x</description><item><title>x</title><link>x</link></item></channel></rss>'
-    WebMock.stub_request(:get, 'http://test.com/feed.rss').to_return(status: 200, body: rss)
-    u = create_user is_admin: true
-    authenticate_with_user(u)
-    query = "query { team(slug: \"#{@team.slug}\") { team_bot_installations(first: 1) { edges { node { smooch_newsletter_information } } } } }"
-    post :create, params: { query: query }
-    assert_response :success
-    assert_not_nil json_response.dig('data', 'team', 'team_bot_installations', 'edges', 0, 'node', 'smooch_newsletter_information')
-  end
-
-  test "should search by similar image on PG" do
-    t = create_team
-    u = create_user is_admin: true
-    authenticate_with_user(u)
-
-    pm = create_project_media team: t
-
-    Bot::Alegre.stubs(:get_items_with_similar_media).returns({ pm.id => 0.8 })
-    path = File.join(Rails.root, 'test', 'data', 'rails.png')
-    file = Rack::Test::UploadedFile.new(path, 'image/png')
-    query = 'query CheckSearch { search(query: "{\"file_type\":\"image\"}") { medias(first: 20) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug, file: file }
-    assert_response :success
-    assert_equal 1, JSON.parse(@response.body)['data']['search']['medias']['edges'].size
-    assert_equal pm.id, JSON.parse(@response.body)['data']['search']['medias']['edges'][0]['node']['dbid']
-    Bot::Alegre.unstub(:get_items_with_similar_media)
-  end
-
-  test "should search by similar image on ES" do
-    setup_elasticsearch
-    t = create_team
-    u = create_user is_admin: true
-    authenticate_with_user(u)
-
-    m = create_claim_media quote: 'Test'
-    m2 = create_claim_media quote: 'Another Test'
-    pm = create_project_media team: t, media: m
-    pm2 = create_project_media team: t, media: m2
-    sleep 2
-
-    Bot::Alegre.stubs(:get_items_with_similar_media).returns({ pm.id => 0.8 })
-    path = File.join(Rails.root, 'test', 'data', 'rails.png')
-    file = Rack::Test::UploadedFile.new(path, 'image/png')
-    query = 'query CheckSearch { search(query: "{\"keyword\":\"Test\",\"file_type\":\"image\"}") { medias(first: 20) { edges { node { dbid } } } } }'
-    post :create, params: { query: query, team: t.slug, file: file }
-    assert_response :success
-    assert_equal 1, JSON.parse(@response.body)['data']['search']['medias']['edges'].size
-    assert_equal pm.id, JSON.parse(@response.body)['data']['search']['medias']['edges'][0]['node']['dbid']
-    Bot::Alegre.unstub(:get_items_with_similar_media)
-  end
-
-  test "should upload search file" do
-    t = create_team
-    u = create_user is_admin: true
-    authenticate_with_user(u)
-
-    path = File.join(Rails.root, 'test', 'data', 'rails.png')
-    file = Rack::Test::UploadedFile.new(path, 'image/png')
-    query = 'mutation { searchUpload(input: {}) { file_handle, file_url } }'
-    post :create, params: { query: query, team: t.slug, file: file }
-    assert_response :success
-    data = JSON.parse(@response.body)['data']['searchUpload']
-    hash = data['file_handle']
-    assert_kind_of String, hash
-    assert CheckS3.exist?("check_search/#{hash}")
-    assert_not_nil data['file_url']
-  end
-
-  test "should get shared teams" do
-    t = create_team
-    f = create_feed
-    f.teams << t
-    query = "query { team(slug: \"#{t.slug}\") { shared_teams } }"
-    post :create, params: { query: query }
-    assert_equal({ t.id.to_s => t.name }, JSON.parse(@response.body).dig('data', 'team', 'shared_teams'))
-    assert_response :success
-  end
-
-  test "should search by feed" do
-    setup_elasticsearch
-    t1 = create_team
-    t2 = create_team
-    u = create_user
-    create_team_user(team: t1, user: u, role: 'editor')
-    authenticate_with_user(u)
-    f = create_feed
-    f.filters = { keyword: 'banana' }
-    f.teams = [t1, t2]
-    f.save!
-
-    # Team 1 content to be shared
-    ft1 = FeedTeam.where(feed: f, team: t1).last
-    ft1.filters = { keyword: 'apple' }
-    ft1.shared = false
-    ft1.save!
-    pm1a = create_project_media quote: 'I like apple and banana', team: t1
-    pm1b = create_project_media quote: 'I like orange and banana', team: t1
-
-    # Team 2 content to be shared
-    ft2 = FeedTeam.where(feed: f, team: t2).last
-    ft2.filters = { keyword: 'orange' }
-    ft2.shared = true
-    ft2.save!
-    pm2a = create_project_media quote: 'I love apple and banana', team: t2
-    pm2b = create_project_media quote: 'I love orange and banana', team: t2
-
-    # Wait for content to be indexed in ElasticSearch
-    sleep 5
-    query = 'query CheckSearch { search(query: "{\"keyword\":\"and\",\"feed_id\":' + f.id.to_s + '}") { medias(first: 20) { edges { node { dbid } } } } }'
-
-    # Can't see anything until content is shared
-    post :create, params: { query: query, team: t1.slug }
-    assert_response :success
-    assert_equal [], JSON.parse(@response.body)['data']['search']['medias']['edges']
-
-    # See content after content is shared
-    with_current_user_and_team(u, t1) do
-      ft1.shared = true
-      ft1.save!
-    end
-    post :create, params: { query: query, team: t1.slug }
-    assert_response :success
-    assert_equal [pm1a.id, pm2b.id].sort, JSON.parse(@response.body)['data']['search']['medias']['edges'].collect{ |pm| pm['node']['dbid'] }.sort
-
-    # Filter by published if feed is published
-    with_current_user_and_team(nil, nil) do
-      f.published = true
-      f.save!
-    end
-    post :create, params: { query: query, team: t1.slug }
-    assert_response :success
-    assert_equal [], JSON.parse(@response.body)['data']['search']['medias']['edges']
-  end
-
-  test "should get feed" do
-    t = create_team private: true
-    create_team_user(user: @u, team: t)
-    f = create_feed
-    query = "query { team(slug: \"#{t.slug}\") { feed(dbid: #{f.id}) { current_feed_team { dbid } } } }"
-
-    post :create, params: { query: query, team: t.slug }
-    assert_nil JSON.parse(@response.body).dig('data', 'team', 'feed')
-
-    with_current_user_and_team(nil, nil) { f.teams << t }
-    post :create, params: { query: query, team: t.slug }
-    assert_equal FeedTeam.where(feed: f, team: t).last.id, JSON.parse(@response.body).dig('data', 'team', 'feed', 'current_feed_team', 'dbid')
-  end
-
-  test "should update feed team" do
-    t1 = create_team private: true
-    create_team_user(user: @u, team: t1, role: 'admin')
-    t2 = create_team private: true
-    f = create_feed
-    f.teams << t1
-    f.teams << t2
-    ft1 = FeedTeam.where(team: t1, feed: f).last
-    ft2 = FeedTeam.where(team: t2, feed: f).last
-    assert !ft1.shared
-    assert !ft2.shared
-
-    query = "mutation { updateFeedTeam(input: { id: \"#{ft1.graphql_id}\", shared: true }) { feed_team { shared } } }"
-    post :create, params: { query: query, team: t1.slug }
-    assert ft1.reload.shared
-
-    query = "mutation { updateFeedTeam(input: { id: \"#{ft2.graphql_id}\", shared: true }) { feed_team { shared } } }"
-    post :create, params: { query: query, team: t2.slug }
-    assert !ft2.reload.shared
   end
 
   protected
